requirements.txt
django-4.2.4 
pillow-10.0.0